Extract validation code to AppleReceiptValidator which conforms to ReceiptValidator protocol
