+++ SwiftyStoreKit/InAppReceipt.swift
@@ -28,12 +28,6 @@ import Foundation
 // MARK - receipt mangement
 internal class InAppReceipt {
 
-    // URL used to verify remotely receipt
-    enum VerifyReceiptURLType: String {
-        case production = "https://buy.itunes.apple.com/verifyReceipt"
-        case sandbox = "https://sandbox.itunes.apple.com/verifyReceipt"
-    }
-
     static var appStoreReceiptUrl: URL? {
         return Bundle.main.appStoreReceiptURL
     }
@@ -70,80 +64,7 @@ internal class InAppReceipt {
                 return
             }
 
-            // Create request
-            let storeURL = URL(string: urlType.rawValue)! // safe (until no more)
-            let storeRequest = NSMutableURLRequest(url: storeURL)
-            storeRequest.httpMethod = "POST"
-
-
-            let requestContents: NSMutableDictionary = [ "receipt-data" : base64EncodedString ]
-            // password if defined
-            if let password = autoRenewPassword {
-                requestContents.setValue(password, forKey: "password")
-            }
-
-            // Encore request body
-            do {
-                storeRequest.httpBody = try JSONSerialization.data(withJSONObject: requestContents, options: [])
-            } catch let e {
-                completion(.error(error: .requestBodyEncodeError(error: e)))
-                return
-            }
-
-            // Remote task
-            let task = session.dataTask(with: storeRequest as URLRequest) { data, response, error -> Void in
-
-                // there is an error
-                if let networkError = error {
-                    completion(.error(error: .networkError(error: networkError)))
-                    return
-                }
-
-                // there is no data
-                guard let safeData = data else {
-                    completion(.error(error: .noRemoteData))
-                    return
-                }
-
-                // cannot decode data
-                guard let receiptInfo = try? JSONSerialization.jsonObject(with: data!, options: .mutableLeaves) as? ReceiptInfo ?? [:] else {
-                    let jsonStr = String(data: safeData, encoding: String.Encoding.utf8)
-                    completion(.error(error: .jsonDecodeError(string: jsonStr)))
-                    return
-                }
-
-                // get status from info
-                if let status = receiptInfo["status"] as? Int {
-                    /*
-                     * http://stackoverflow.com/questions/16187231/how-do-i-know-if-an-in-app-purchase-receipt-comes-from-the-sandbox
-                     * How do I verify my receipt (iOS)?
-                     * Always verify your receipt first with the production URL; proceed to verify
-                     * with the sandbox URL if you receive a 21007 status code. Following this
-                     * approach ensures that you do not have to switch between URLs while your
-                     * application is being tested or reviewed in the sandbox or is live in the
-                     * App Store.
-
-                     * Note: The 21007 status code indicates that this receipt is a sandbox receipt,
-                     * but it was sent to the production service for verification.
-                     */
-                    let receiptStatus = ReceiptStatus(rawValue: status) ?? ReceiptStatus.unknown
-                    if case .testReceipt = receiptStatus {
-                        verify(urlType: .sandbox, password: autoRenewPassword, session: session, completion: completion)
-                    }
-                    else {
-                        if receiptStatus.isValid {
-                            completion(.success(receipt: receiptInfo))
-                        }
-                        else {
-                            completion(.error(error: .receiptInvalid(receipt: receiptInfo, status: receiptStatus)))
-                        }
-                    }
-                }
-                else {
-                    completion(.error(error: .receiptInvalid(receipt: receiptInfo, status: ReceiptStatus.none)))
-                }
-            }
-            task.resume()
+			validator.validate(receipt: base64EncodedString, password: autoRenewPassword, completion: completion)
     }
   
     /**
